Merge CvParagraph into CvListItem component
CvParagraph was being used as list items. In this app, both components are quite similar; the main differences are the absence of an icon and the presence of spacing below for CvParagraph.

To simplify the application, CvListItem is now used in place of CvParagraph. Enhancements were made to CvListItem to allow disabling the icon and adding spacing below when needed.