Increase delay for displaying preview to 1000 ms
The preview was previously delayed by 600 ms to avoid a BindingError in react-pdf. The error happened because of a bug in the react-pdf library when displaying the preview and download button at the same time. However, the error still occurred in some cases, so the delay was increased to 1000 ms.